svd: full device parsing (big monolith commit currently, might split)
This is a massive expansion of svada to support a lot of new features
not previously supported.

TODO: FINISH

The API for parsing an SVD file has been changed from
"parse_peripheral()" to just "parse()", and now returns a
Device instance instead of a Peripheral.

The XML parser backend was changed to lxml to make use of the objectify
API that maps XML elements to Python objects. A "binding" API was
written on top of the lxml/objectify API to make accessing the XML
elements/attributes convenient and to minimize the boilerplate
associated with reading these properties.

Ref: NCSDK-21485

Signed-off-by: Jonathan Nilsen <jonathan.nilsen@nordicsemi.no>